Rename `csympy` --> `symengine`
