+++ symengine/lib/symengine.pxd
